+++ tests/dummy/app/JsonApi/V1/Posts/PostResource.php
@@ -55,7 +55,7 @@ class PostResource extends JsonApiResource
         return [
             $this->relation('author')->showDataIfLoaded(),
             $this->relation('comments'),
-            $this->relation('tags')->showDataIfLoaded()->mustValidate(),
+            $this->relation('tags')->showDataIfLoaded(),
         ];
     }