+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Changed
+- [#2](https://github.com/laravel-json-api/laravel/issues/2) **BREAKING** Improved the extraction of existing resource
+  field values when constructing validation data for update requests:
+    - The `existingAttributes()` and `existingRelationships()` methods on the resource request class has been removed.
+      If you need to modify the existing values before the client values are merged, implement the `withExisting()`
+      method instead. This receives the model its JSON representation (as an array).
+    - The `mustValidate()` method must now be called on a schema relationship field. (Previously this was on the
+      resource relation.) By default, belongs-to and morph-to relations will be included when extracting existing
+      values; all other relations will not. Use the `mustValidate()` or `notValidated()` method on the schema relation
+      to alter whether a relation is included in the extracted values.
+
 ## [1.0.0-alpha.2] - 2021-02-02
 
 ### Added